Add json output
